+++ server/bin/esm.js
@@ -2,4 +2,4 @@
 
 // https://github.com/standard-things/esm
 require = require('esm')(module)
-module.exports = require('../server.js').default
+module.exports = require('../src/server.js').default